requirements.txt
requests
nbformat
python-dotenv

pandas